replace old `add` with updated form version
also fix key rendering of form
could also be doing this form by writing WET forms for this minimal usage
